% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{amd_order}
\alias{amd_order}
\title{AMD ordering}
\usage{
amd_order(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
A list with the permutation vector and it's inverse.
}
\description{
AMD ordering
}
\details{
Generates the approximate minimum degree ordering of the matrix for use in efficient
Cholesky decomposition of PAP^T.
}
