% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{LDL_Cholesky}
\alias{LDL_Cholesky}
\title{Generate LDL decomposition from Matrix class `dsCMatrix`}
\usage{
LDL_Cholesky(mat)
}
\arguments{
\item{mat}{A matrix of class `dsCMatrix`}
}
\value{
A list of matrices L and D
}
\description{
Generates the Cholesky decomposition L as A == LL^T from a sparse 
matrix
}
