% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{LL_Cholesky}
\alias{LL_Cholesky}
\title{Generate Cholesky decomposition from Matrix class `dsCMatrix`}
\usage{
LL_Cholesky(mat)
}
\arguments{
\item{mat}{A matrix of class `dsCMatrix`}
}
\value{
A matrix of class `ddiMatrix`
}
\description{
Generates the Cholesky decomposition L as A == LL^T from a 
sparse matrix
}
