% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{sparse_D}
\alias{sparse_D}
\title{Generate matrix D from `sparse_chol` output}
\usage{
sparse_D(mat)
}
\arguments{
\item{mat}{List returned by `sparse_chol`}
}
\value{
A matrix of class `ddiMatrix`
}
\description{
Generates the D matrix of the LDL decomposition from the output of 
the `sparse_chol` function
}
