% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{sparse_L}
\alias{sparse_L}
\title{Generate matrix L from `sparse_chol` output}
\usage{
sparse_L(mat)
}
\arguments{
\item{mat}{List returned by `sparse_chol`}
}
\value{
A matrix of class `dsCMatrix`
}
\description{
Generates the L matrix of the LDL decomposition from the output of 
the `sparse_chol` function
}
