\name{SparseChol-package}
\alias{SparseChol-package}
\alias{SparseChol}
\docType{package}
\title{\packageTitle{SparseChol}}
\description{\packageDescription{SparseChol}}
\section{Package Content}{\packageIndices{SparseChol}}
\author{\packageAuthor{SparseChol}}
\section{Maintainer}{\packageMaintainer{SparseChol}}
\keyword{package}
