% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sparse_chol}
\alias{sparse_chol}
\title{Sparse Cholesky decomposition}
\usage{
sparse_chol(n, Ap, Ai, Ax)
}
\arguments{
\item{n}{Integer specifying the dimension of the matrix}

\item{Ap}{numeric (integer valued) vector of pointers, one for each column (or row), to the initial (zero-based) index of elements in the column (or row).}

\item{Ai}{Integer vector specifying the row positions of the non-zero values of the matrix}

\item{Ax}{values of the non-zero matrix entries}
}
\value{
A list with elements n, Ai, Ap, Ax (corresponding to above arguments) for matrix L, and element D, which 
contains the diagonal values of matrix D.
}
\description{
Sparse Cholesky decomposition
}
\details{
Generates the LDL decomposition of a symmetrix, sparse matrix using the method 
described by Timothy Davis (see references). Required input is a matrix
in sparse format from the matrix package, see \link[Matrix]{sparseMatrix}
}
